+++ ahnlich/types/src/keyval.rs
@@ -20,10 +20,10 @@ impl fmt::Display for StoreName {
 /// A store value for now is a simple key value pair of strings
 pub type StoreValue = StdHashMap<MetadataKey, MetadataValue>;
 
-/// A store key is always an f64 one dimensional array
+/// A store key is always an f32 one dimensional array
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(transparent)]
-pub struct StoreKey(pub Array1<f64>);
+pub struct StoreKey(pub Array1<f32>);
 
 impl StoreKey {
     pub fn dimension(&self) -> usize {
@@ -38,14 +38,14 @@ impl PartialEq for StoreKey {
         if self.0.shape() != other.0.shape() {
             return false;
         }
-        // std::f64::EPSILON adheres to the IEEE 754 standard and we use it here to determine when
-        // two Array1<f64> are extremely similar to the point where the differences are neglible.
+        // std::f32::EPSILON adheres to the IEEE 754 standard and we use it here to determine when
+        // two Array1<f32> are extremely similar to the point where the differences are neglible.
         // We can modify to allow for greater precision, however we currently only
         // use it for PartialEq and not for it's distinctive properties. For that, within the
-        // server we defer to using StoreKeyId whenever we want to compare distinctive Array1<f64>
+        // server we defer to using StoreKeyId whenever we want to compare distinctive Array1<f32>
         self.0
             .iter()
             .zip(other.0.iter())
-            .all(|(x, y)| (x - y).abs() < std::f64::EPSILON)
+            .all(|(x, y)| (x - y).abs() < std::f32::EPSILON)
     }
 }
